IBX-3114: Add Pagination helper
